news-summarizer-rag/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
schedule==1.2.0
tqdm==4.66.1
colorama==0.4.6
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2